src/requirements.txt
pyyaml
requests
python-dotenv==1.0.1
psycopg2